+++ NEWS.md
@@ -1,5 +1,9 @@
 ## ironseed (development)
 
+* Expand `ironseed()` to support multiple methods out of the box for collecting
+  seeds.
+  - "args" method extracts one or more seeds from the command line.
+  - "env" method extracts a seed from an environment variable named "IRONSEED".
 * Better support for MacOS and PowerPC (from Sergey Fedorov @barracuda156)
 * Better support for \*BSD systems
 * Better support for big-endian systems